+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2022-04-28
+
+### Changed
+
+- Object ProductPrices:
+  - Add field termCode and periodicityCode
+
 ## [3.3.0] - 2022-04-27
 
 ### Changed